Split oversized validator registry tests
