+++ .pm/tasks/task_f9b662cdbb354c28a431aa66ebd0848f.execution.md
@@ -201,3 +201,12 @@ Example:
 - Blocker / Next Action: Continue `normal_pr_ci_watch`: monitor GitHub checks, mergeability, PR comments, and review threads; fix any failing checks or actionable comments before merge.
 - PR Purpose Decision: normal_pr_ci_watch
 - PR URL: https://github.com/eng-cc/oasis7/pull/357
+
+## 2026-06-04 20:45:50 CST / tpm
+- 完成内容: Addressed the PR #357 Rust CI file-size failure.
+- 遗留事项: GitHub PR checks must be rerun on the pushed fix before merge; real four-node deployment/restart validation remains post-PR operational work.
+- Action: Split the new stake-weighted governance finality test into `governance_finality_tests.rs` and moved the stake-weighted module-release attestation test into the existing release-controls child test file, keeping behavior unchanged while bringing the oversized parent test files below the 1200-line gate.
+- Validation Command: `wc -l crates/oasis7/src/runtime/tests/governance.rs crates/oasis7/src/runtime/tests/governance_finality_tests.rs crates/oasis7/src/runtime/tests/module_action_loop_release_controls.rs crates/oasis7/src/runtime/tests/module_action_loop_release_controls_tests.rs`; `./scripts/check-rust-file-size.sh`; `./scripts/cargo-dev.sh test -p oasis7 governance_apply_with_finality_uses_stake_weighted_threshold_bps -- --nocapture`; `./scripts/cargo-dev.sh test -p oasis7 module_release_apply_uses_stake_weighted_attestation_threshold_bps -- --nocapture`; `./scripts/cargo-dev.sh check -p oasis7`; `bash -n scripts/p2p-triad-node-start.sh && bash -n scripts/p2p-public-testnet-local-observer-sync.sh && ./scripts/doc-governance-check.sh && git diff --check`
+- Expected Result: No Rust files exceed the configured size limit; the two moved focused tests are still discovered and pass.
+- Actual Result: `governance.rs` is 1168 lines, `governance_finality_tests.rs` is 85 lines, `module_action_loop_release_controls.rs` is 1161 lines, and `module_action_loop_release_controls_tests.rs` is 1071 lines. `check-rust-file-size: OK`. Both focused tests passed 1/1. `check -p oasis7` passed with only existing launcher/faucet warnings. Script syntax checks passed, `doc-governance-check: OK`, and `git diff --check` passed.
+- Blocker / Next Action: Run final local sanity checks, commit/push the CI fix, then continue normal PR CI watch.
